backend/requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
bcrypt==4.1.2

# CORS
Flask-CORS==4.0.0

# WebSocket
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0

# Redis (for caching and WebSocket message queue)
redis==5.0.1
Flask-Caching==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Email
Flask-Mail==0.9.1

# Validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Rate Limiting
Flask-Limiter==3.5.0

# Date/Time
python-dateutil==2.8.2

# File Handling
Pillow==10.1.0

# Utilities
requests==2.31.0
gunicorn==21.2.0  # For production deployment

# Development Tools
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
